src/main/java/by/bsu/wialontransport/controller/initbinder/InitBinderControllerAdvice.java
package by.bsu.wialontransport.controller.initbinder;

import org.springframework.beans.propertyeditors.StringTrimmerEditor;
import org.springframework.web.bind.WebDataBinder;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.InitBinder;

@ControllerAdvice
public final class InitBinderControllerAdvice {

    @InitBinder
    public void configureInitBinder(final WebDataBinder webDataBinder) {
        //тримит строки при binding-е, если были введены одни пробелы, то будет присвоено полю значиение null
        final StringTrimmerEditor stringTrimmerEditor = new StringTrimmerEditor(true);
        webDataBinder.registerCustomEditor(String.class, stringTrimmerEditor);
    }

}
